requirements/base.txt
# Core runtime dependencies for the root trading bot (main.py)
ccxt>=4.0,<5.0
numpy>=1.26,<3.0
pandas>=2.2,<4.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
